+++ cmd/zed/index/convert/command.go
@@ -11,7 +11,7 @@ import (
 	"github.com/brimdata/zed/pkg/charm"
 	"github.com/brimdata/zed/pkg/iosrc"
 	"github.com/brimdata/zed/zbuf"
-	"github.com/brimdata/zed/zio/detector"
+	"github.com/brimdata/zed/zio/anyio"
 	"github.com/brimdata/zed/zson"
 )
 
@@ -73,7 +73,7 @@ func (c *Command) Run(args []string) error {
 		path = iosrc.Stdin
 	}
 	zctx := zson.NewContext()
-	file, err := detector.OpenFile(zctx, path, c.inputFlags.Options())
+	file, err := anyio.OpenFile(zctx, path, c.inputFlags.Options())
 	if err != nil {
 		return err
 	}
