+++ zio/anyio/file_test.go
@@ -1,4 +1,4 @@
-package detector
+package anyio
 
 import (
 	"context"
