+++ zio/emitter/file.go
@@ -10,7 +10,7 @@ import (
 	"github.com/brimdata/zed/pkg/terminal"
 	"github.com/brimdata/zed/zbuf"
 	"github.com/brimdata/zed/zio"
-	"github.com/brimdata/zed/zio/detector"
+	"github.com/brimdata/zed/zio/anyio"
 )
 
 func NewFile(ctx context.Context, path string, opts zio.WriterOpts) (zbuf.WriteCloser, error) {
@@ -56,7 +56,7 @@ func NewFileWithSource(ctx context.Context, path iosrc.URI, opts zio.WriterOpts,
 	// On close, zbuf.WriteCloser.Close() will close and flush the
 	// downstream writer, which will flush the bufwriter here and,
 	// in turn, close its underlying writer.
-	w, err := detector.LookupWriter(wc, opts)
+	w, err := anyio.LookupWriter(wc, opts)
 	if err != nil {
 		return nil, err
 	}
